+++ src/networking/modules/react_native.js
@@ -1,111 +1,48 @@
 /* @flow */
+/* global FormData */
+/* global fetch */
 
-import {
-  EndpointDefinition,
-  StatusAnnouncement,
-} from '../../core/flow_interfaces';
-import { buildUrl } from '../utils';
+import superagent from 'superagent';
+import { EndpointDefinition } from '../../core/flow_interfaces';
+import { xdr, postfile as postfilewebnode } from './web-node';
 
-declare var fetch: any;
-
-function log(url, qs, res) {
-  let _pickLogger = () => {
-    if (console && console.log) return console; // eslint-disable-line no-console
-    return console;
-  };
-
-  let start = new Date().getTime();
-  let timestamp = new Date().toISOString();
-  let logger = _pickLogger();
-  logger.log('<<<<<'); // eslint-disable-line no-console
-  logger.log(`[${timestamp}]`, '\n', url, '\n', qs); // eslint-disable-line no-console
-  logger.log('-----'); // eslint-disable-line no-console
-
-  let now = new Date().getTime();
-  let elapsed = now - start;
-  let timestampDone = new Date().toISOString();
-
-  logger.log('>>>>>>'); // eslint-disable-line no-console
-  logger.log(`[${timestampDone} / ${elapsed}]`, '\n', url, '\n', qs, '\n', res); // eslint-disable-line no-console
-  logger.log('-----');
-}
-
-function xdr(
-  method: string,
+async function postfileuri(
   url: string,
-  params: Object,
-  body: string,
-  endpoint: EndpointDefinition,
-  callback: Function
-): void {
-  let status: StatusAnnouncement = {};
-  status.operation = endpoint.operation;
+  fields: $ReadOnlyArray<{ key: string, value: string }>,
+  fileInput: any
+): Promise<any> {
+  const formData = new FormData();
 
-  fetch(buildUrl(url, params), { method, body })
-    .then((response) => {
-      status.error = false;
+  fields.forEach(({ key, value }) => {
+    formData.append(key, value);
+  });
 
-      if (response.status) {
-        status.statusCode = response.status;
-      }
+  formData.append('file', fileInput);
 
-      return response.json();
-    })
-    .then((response) => {
-      let resp = response;
+  const result = await fetch(url, {
+    method: 'POST',
+    body: formData
+  });
 
-      if (this._config.logVerbosity) {
-        log(url, params, resp);
-      }
-
-      callback(status, resp);
-    })
-    .catch((e) => {
-      status.error = true;
-      status.errorData = e.error;
-      status.category = this._detectErrorCategory(e.error);
-      callback(status, null);
-    });
+  return result;
 }
 
-export function get(
-  params: Object,
-  endpoint: EndpointDefinition,
-  callback: Function
-) {
-  let url = this.getStandardOrigin() + endpoint.url;
-
-  return xdr.call(this, 'GET', url, params, '', endpoint, callback);
+export async function postfile(
+  url: string,
+  fields: $ReadOnlyArray<{ key: string, value: string }>,
+  fileInput: any
+): Promise<any> {
+  if (!fileInput.uri) {
+    return postfilewebnode(url, fields, fileInput);
+  } else {
+    return postfileuri(url, fields, fileInput);
+  }
 }
 
-export function post(
-  params: Object,
-  body: string,
-  endpoint: EndpointDefinition,
-  callback: Function
-) {
-  let url = this.getStandardOrigin() + endpoint.url;
-
-  return xdr.call(this, 'POST', url, params, body, endpoint, callback);
-}
-
-export function patch(
-  params: Object,
-  body: string,
-  endpoint: EndpointDefinition,
-  callback: Function
-) {
-  let url = this.getStandardOrigin() + endpoint.url;
-
-  return xdr.call(this, 'PATCH', url, params, body, endpoint, callback);
-}
-
-export function del(
-  params: Object,
-  endpoint: EndpointDefinition,
-  callback: Function
-) {
-  let url = this.getStandardOrigin() + endpoint.url;
-
-  return xdr.call(this, 'DELETE', url, params, '', endpoint, callback);
+export function getfile(params: Object, endpoint: EndpointDefinition, callback: Function): Promise<any> {
+  let superagentConstruct = superagent
+    .get(this.getStandardOrigin() + endpoint.url)
+    .set(endpoint.headers)
+    .query(params);
+  return xdr.call(this, superagentConstruct, endpoint, callback);
 }