+++ src/react_native/index.js
@@ -5,11 +5,12 @@ import PubNubCore from '../core/pubnub-common';
 import Networking from '../networking';
 import Database from '../db/common';
 import Cbor from '../cbor/common';
-import { del, get, post, patch, file } from '../networking/modules/web-node';
+import { del, get, post, patch } from '../networking/modules/web-node';
+import { getfile, postfile } from '../networking/modules/react_native';
 import { InternalSetupStruct } from '../core/flow_interfaces';
 
 import WebCryptography from '../crypto/modules/web';
-import PubNubFile from '../file/modules/web';
+import PubNubFile from '../file/modules/react-native';
 
 export default class extends PubNubCore {
   constructor(setup: InternalSetupStruct) {
@@ -19,7 +20,7 @@ export default class extends PubNubCore {
     setup.PubNubFile = PubNubFile;
     setup.cryptography = new WebCryptography();
 
-    setup.networking = new Networking({ del, get, post, patch, file });
+    setup.networking = new Networking({ del, get, post, patch, getfile, postfile });
     setup.sdkFamily = 'ReactNative';
     setup.ssl = true;